portal-impl: Rename files to follow the prefix-name rule
Renames the portal-impl files to xdp-portal-impl.
