+++ src/xdg-desktop-portal.c
@@ -34,6 +34,7 @@
 #include "xdp-dbus.h"
 #include "xdp-impl-dbus.h"
 #include "xdp-method-info.h"
+#include "xdp-portal-impl.h"
 #include "xdp-session-persistence.h"
 
 #include "account.h"
@@ -55,7 +56,6 @@
 #include "notification.h"
 #include "open-uri.h"
 #include "permissions.h"
-#include "portal-impl.h"
 #include "power-profile-monitor.h"
 #include "print.h"
 #include "proxy-resolver.h"
