+++ src/xdp-portal-impl.c
@@ -23,7 +23,7 @@
 
 #include "config.h"
 
-#include "portal-impl.h"
+#include "xdp-portal-impl.h"
 
 #include <stdio.h>
 #include <string.h>
